requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1;python_version<"3.9"
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cron-descriptor==1.2.35
Django==4.1.4
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-pgschemas==0.12.0
django-recurrence==1.11.1
django-tables2==2.5.0
django-timezone-field==5.0
django-utils-six==2.0
django-widget-tweaks==1.4.12
gunicorn==20.1.0
kombu==5.2.4
ordered-set==4.1.0
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2==2.9.5
psycopg2-binary==2.9.6
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.5.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
vine==5.0.0
wcwidth==0.2.6
